Creato APIApplication.java per avere un path assoluto per le API Rest; eliminati i serializable in tutte le classi dove non era necessario; modificati tutti gli id del programma da int a long; aggiunta gestione delle sequence negli oggetti di persistenza; specificato lo schema negli oggetti di persistenza; specificate tutte le colonne negli oggetti di persistenza; inserito parte del plugin delle mapstruct mancante che ne impediva la corretta generazione; scritta l'implementazione di tutti i dao services; sistemato con gli embedded la coppia di ID di FatturaArticoloEntity.java; aggiunte annotazioni di lombok ToString con relative esclusioni per evitare loop infiniti quando si scaricano i dati dal DB; sistemato leggermente il persistence.xml; ridotta significativamente la lunghezza del pom utilizzando quasi solo le librerie native di WildFly; sistemata e migliorata la procedura di generazione delle fatture casuali.

+++ src/main/java/com/advancia/OfficineMilanesi/application/api/rest/ClienteRestController.java
@@ -24,7 +24,7 @@ public class ClienteRestController {
 
     @GET
     @Path("/{id}")
-    public ClienteDto getClienteById(@PathParam("id") int id) {
+    public ClienteDto getClienteById(@PathParam("id") long id) {
         ClienteDto clienteDto = ClienteService.getClienteById(id);
         if(clienteDto == null) throw new NotFoundException("Cliente con id " + id + " non trovato.");
         return clienteDto;
@@ -38,14 +38,14 @@ public class ClienteRestController {
 
     @PUT
     @Path("/{id}")
-    public Response updateCliente(@PathParam("id") int id, ClienteDto clienteAgg) {
+    public Response updateCliente(@PathParam("id") long id, ClienteDto clienteAgg) {
         ClienteService.updateCliente(id, clienteAgg);
         return Response.ok().build();
     }
 
     @DELETE
     @Path("/{id}")
-    public Response deleteCliente(@PathParam("id") int id) {
+    public Response deleteCliente(@PathParam("id") long id) {
         ClienteService.deleteCliente(id);
         return Response.noContent().build();
     }
